+++ src/Service/Database/DatabaseFactory.php
@@ -6,6 +6,7 @@ use Kuusamo\Vle\Helper\Environment;
 
 use Doctrine\Common\EventManager;
 use Doctrine\ORM\Tools\Setup;
+use Doctrine\ORM\Configuration;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Events;
 
@@ -18,10 +19,6 @@ class DatabaseFactory
      */
     public static function create()
     {
-        $isDevMode = true;
-        $paths = [__DIR__ . '/../../Entity'];
-        $metadata = Setup::createAnnotationMetadataConfiguration($paths, $isDevMode);
-
         // database configuration parameters
         $dbParams = [
             'driver'   => Environment::get('DB_DRIVER', 'pdo_mysql'),
@@ -32,12 +29,30 @@ class DatabaseFactory
             'charset'  => 'utf8'
         ];
 
+        // doctrine configuration
+        $devMode = Environment::get('ENVIRONMENT') == 'development';
+
+        $config = new Configuration;
+
+        $driverImpl = $config->newDefaultAnnotationDriver(
+            __DIR__ . '/../../Entity'
+        );
+
+        $config->setProxyDir(__DIR__ . '/../../Proxy');
+        $config->setProxyNamespace('Kuusamo\Vle\Proxy');
+        $config->setMetadataDriverImpl($driverImpl);
+        $config->setAutoGenerateProxyClasses($devMode);
+
+        $cache = new \Doctrine\Common\Cache\ArrayCache;
+        $config->setMetadataCacheImpl($cache);
+        $config->setQueryCacheImpl($cache);
+
         // event manager and table prefix
         $evm = new EventManager;
         $tablePrefix = new TablePrefix(Environment::get('DB_PREFIX', 'vle_'));
         $evm->addEventListener(Events::loadClassMetadata, $tablePrefix);
         
         // obtaining the entity manager
-        return EntityManager::create($dbParams, $metadata, $evm);
+        return EntityManager::create($dbParams, $config, $evm);
     }
 }